+++ Changelog.md
@@ -3,6 +3,8 @@
 - Added Domination gamemode. Teams fight to control as many maps as possible before time runs out.
 - Added Message of the Day section to the bingo main menu. This allows for quick informing of updates, server changes and PSAs for known issues.
 - Added special in-game ranks obtainable by being a noteworthy member of the larger ULTRAKILL community (mapper, modder, speedrunner, etc), or by having supported the mod on Kofi. If you fit any of the criteria, feel free to reach out to me to recieve your in-game role!
+- Added the ability to right click on a level in the bingo card on the pause menu, to ping the level for the rest of your team.
+- Changed text on the pause menu to better indicate possible actions (left click to move to level, right click to ping map, etc.)
 - When a player (re)claims a level, the new time requirement is now displayed in the HUD message that's broadcasted to all players.
 - Visual pass to make UI elements more in line with UI changes from the latest game update.